+++ sharding/config/config.go
@@ -2,19 +2,9 @@ package config
 
 import (
 	"fmt"
-	"net/url"
-	"strings"
 
-	"errors"
-
-	units "github.com/docker/go-units"
+	"github.com/allegro/akubra/types"
 )
-
-// ClusterConfig defines cluster configuration
-type ClusterConfig struct {
-	// Backends should contain s3 backend urls
-	Backends []YAMLUrl `yaml:"Backends"`
-}
 
 // MultiClusterConfig defines region settings for multicluster
 type MultiClusterConfig struct {
@@ -34,36 +24,12 @@ type RegionConfig struct {
 	Default bool `yaml:"Default,omitempty"`
 }
 
-// YAMLUrl type fields in yaml configuration will parse urls
-type YAMLUrl struct {
-	*url.URL
-}
+// YAMLUrl is yaml deserializing wrapper for net/url.URL
+type YAMLUrl = types.YAMLUrl
 
 // SyncLogMethod type fields in yaml configuration will parse list of HTTP methods
 type SyncLogMethod struct {
 	Method string
-}
-
-// AdditionalHeaders type fields in yaml configuration will parse list of special headers
-type AdditionalHeaders map[string]string
-
-// HumanSizeUnits type for max. payload body size in bytes
-type HumanSizeUnits struct {
-	SizeInBytes int64
-}
-
-// UnmarshalYAML for YAMLUrl
-func (yurl *YAMLUrl) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var s string
-	if err := unmarshal(&s); err != nil {
-		return err
-	}
-	url, err := url.Parse(s)
-	if url.Host == "" {
-		return fmt.Errorf("url should match proto://host[:port]/path scheme - got %q", s)
-	}
-	yurl.URL = url
-	return err
 }
 
 // UnmarshalYAML for SyncLogMethod
@@ -83,39 +49,4 @@ func (slm *SyncLogMethod) UnmarshalYAML(unmarshal func(interface{}) error) error
 	return nil
 }
 
-// UnmarshalYAML for AdditionalHeaders
-func (ah *AdditionalHeaders) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var headers map[string]string
-	if err := unmarshal(&headers); err != nil {
-		return err
-	}
-	for key, value := range headers {
-
-		if strings.TrimSpace(key) == "" {
-			return fmt.Errorf("Empty additional header with value: %q", value)
-		}
-		if strings.TrimSpace(value) == "" {
-			return fmt.Errorf("Empty additional header with key: %q", key)
-		}
-
-	}
-	*ah = AdditionalHeaders(headers)
-	return nil
-}
-
-// UnmarshalYAML for HumanSizeUnits
-func (hsu *HumanSizeUnits) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var size string
-	if err := unmarshal(&size); err != nil {
-		return err
-	}
-	value, err := units.FromHumanSize(size)
-	if err != nil {
-		return fmt.Errorf("Unable to parse BodyMaxSize: %s" + err.Error())
-	}
-	if value < 1 {
-		return errors.New("BodyMaxSize must be greater than zero")
-	}
-	hsu.SizeInBytes = value
-	return nil
-}
+type HumanSizeUnits = types.HumanSizeUnits
